+++ com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/HDProfileId.cs
@@ -193,6 +193,7 @@ namespace UnityEngine.Rendering.HighDefinition
         DepthOfFieldGatherNear,
         DepthOfFieldPreCombine,
         DepthOfFieldCombine,
+        LensFlare,
         MotionBlur,
         MotionBlurMotionVecPrep,
         MotionBlurTileMinMax,